pyproject.toml
[tool.poetry]
name = "failure-prediction"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
scikit-learn = "^1.5.2"
streamlit = "^1.38.0"
openai = "^1.50.1"
matplotlib = "^3.9.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
